+++ packages/starboard-notebook/src/components/helpers/markdown.ts
@@ -1,6 +1,5 @@
 import mdlib from "markdown-it";
 import { hookMarkdownItCrossOriginImages } from "./crossOriginIsolated";
-import { hookMarkdownItToEmojiPlugin } from "./emoji";
 import { hookMarkdownItToCodemirrorHighlighter } from "./highlight";
 import { hookMarkdownItToKaTeX } from "./katex";
 
@@ -9,7 +8,6 @@ export function getMarkdownItWithDefaultPlugins(
 ): { md: mdlib; katexLoaded: Promise<void> } {
   const md = new mdlib(markdownitOpts);
   hookMarkdownItToCodemirrorHighlighter(md);
-  hookMarkdownItToEmojiPlugin(md);
   hookMarkdownItCrossOriginImages(md);
   const prom = hookMarkdownItToKaTeX(md);