+++ packages/starboard-rich-editor/src/rich-editor/lib/markdown/math.ts
@@ -1,14 +1,11 @@
 // @ts-nocheck
 // Based on https://github.com/iktakahiro/markdown-it-katex/blob/master/index.js (MIT license)
+
 /* Process inline math */
 /*
 Like markdown-it-simplemath, this is a stripped down, simplified version of:
 https://github.com/runarberg/markdown-it-math
-It differs in that it takes (a subset of) LaTeX as input and relies on KaTeX
-for rendering output.
 */
-
-import katex from "katex";
 
 // Test if potential opening or closing delimieter
 // Assumes that there is a "$" at state.src[pos]
@@ -116,7 +113,7 @@ function math_inline(state, silent) {
   return true;
 }
 
-function math_block(state, start, end, silent) {
+function math_display(state, start, end, silent) {
   var firstLine,
     lastLine,
     next,
@@ -169,67 +166,21 @@ function math_block(state, start, end, silent) {
 
   state.line = next + 1;
 
-  token = state.push("math_block", "math", 0);
+  token = state.push("math_display", "math", 0);
   token.block = true;
   token.content =
     (firstLine && firstLine.trim() ? firstLine + "\n" : "") +
     state.getLines(start + 1, next, state.tShift[start], true) +
     (lastLine && lastLine.trim() ? lastLine : "");
+
   token.map = [start, state.line];
   token.markup = "$$";
   return true;
 }
 
-function escapeHtml(unsafe) {
-  return unsafe
-    .replace(/&/g, "&amp;")
-    .replace(/</g, "&lt;")
-    .replace(/>/g, "&gt;")
-    .replace(/"/g, "&quot;")
-    .replace(/'/g, "&#039;");
-}
-
 export default function (md, options) {
-  // Default options
-  options = options || {};
-
-  // set KaTeX as the renderer for markdown-it-simplemath
-  var katexInline = function (latex) {
-    options.displayMode = false;
-    try {
-      return katex.renderToString(latex, options);
-    } catch (error) {
-      if (options.throwOnError) {
-        console.log(error);
-      }
-      return `<span class='katex-error' title='${escapeHtml(error.toString())}'>${escapeHtml(latex)}</span>`;
-    }
-  };
-
-  var inlineRenderer = function (tokens, idx) {
-    return katexInline(tokens[idx].content);
-  };
-
-  var katexBlock = function (latex) {
-    options.displayMode = true;
-    try {
-      return "<p class='katex-block'>" + katex.renderToString(latex, options) + "</p>";
-    } catch (error) {
-      if (options.throwOnError) {
-        console.log(error);
-      }
-      return `<p class='katex-block katex-error' title='${escapeHtml(error.toString())}'>${escapeHtml(latex)}</p>`;
-    }
-  };
-
-  var blockRenderer = function (tokens, idx) {
-    return katexBlock(tokens[idx].content) + "\n";
-  };
-
   md.inline.ruler.after("escape", "math_inline", math_inline);
-  md.block.ruler.after("blockquote", "math_block", math_block, {
+  md.block.ruler.after("blockquote", "math_display", math_display, {
     alt: ["paragraph", "reference", "blockquote", "list"],
   });
-  md.renderer.rules.math_inline = inlineRenderer;
-  md.renderer.rules.math_block = blockRenderer;
 }